Music_Player
This application reads internal memory of a device and fetches all sort of audio files .
The user gets to see those files in a listview and can choose a file to play .
The user can control these files using few buttons like : play/pause , forward, rewind, next, previous.
Seekbar is used to show progress of the audio and also to control the volume.